requirements.txt
numpy
numpy-stl
transforms3d
pygltflib
networkx
matplotlib
mujoco>=2.3.0
jax>=0.4.0
jaxlib>=0.4.0
flax>=0.7.0
ml_collections>=0.1.0
etils>=0.9.0
orbax-checkpoint>=0.4.0
mujoco-mjx
tensorboard
gymnasium
imageio
stable_baselines3
sbx-rl